Update StyleMeProject.py
This is the initial code for the StyleMe quiz! Originally, I wanted to utilize a pre-established code, but after attempting to create the code I need, it would be unnecessarily long because I would have to do each genre individually. At my level of understanding the code would not connect and flow in a way that would allow it to run successfully. Now I am utilizing a different format to code for this assignment.