Update StyleMeProject.py
I just deleted my whole code and added in a new and improved version. After the initial code was written, I had quite a few errors. In order to properly work through all of them, I utilized Jupyter notebook. The main errors were unnecessary spaces, missing commas in the artist list, wrong indentation, misspelled words, and missing characters/symbols. For example, I was getting a tuple error on each one of my artist list because I did not have a comma after every parenthesis was closed. Instead of artist, I previously wrote artist, so python did not recognize it. There was a missing \ in front of a statement, so it was not recognized. As well as my indentation after my if statement was not right. Now I have been able to get the code to run, except the actual fashion description is not added yet. As well as, I am having an issue with the introductory statement not appearing until the end of my code.