Update StyleMeProject.py
In this update I focused on created the actual fashion descriptions and then implementing them so that they would go with the right era for each genre. This was quite difficult and took a while but figured it out in the end. I used around 3 different methods trying to figure out how to do this.